feat(workspace-cap): port repository + WorkspaceError (slice 5/N)
ADR-152 §3 F3.4 slice 5/N — verbatim port of repository.rs (705 lines,
PostgreSQL persistence) plus extraction of WorkspaceError into the
dasclaw_workspace_cap crate. Continues #696 / #698.

1. Extract WorkspaceError (12-variant thiserror enum, 0 internal deps)
   from desktop-client/ironclaw/src/error.rs into
   crates/dasclaw_workspace_cap/src/error.rs. Original site becomes a
   pub-use re-export shim so Error::Workspace(#[from] WorkspaceError)
   keeps working unchanged.

2. git mv desktop-client/ironclaw/src/workspace/repository.rs into
   crates/dasclaw_workspace_cap/src/repository.rs. Function bodies
   verbatim. Only three import-path adjustments:
   - crate::workspace::document::*  -> crate::document::*
   - crate::workspace::search::*    -> crate::search::*
   - crate::workspace::merge_workspace_entries(..) ->
     crate::document::merge_workspace_entries(..)   (function actually
     defined in document.rs:152; ironclaw used to reach it through the
     workspace/mod.rs re-export)
   Old path becomes a 6-line pub-use shim. mod.rs:51 keeps its existing
   #[cfg(feature = "postgres")] mod repository; gate.

3. cap Cargo.toml: add 4 optional deps (deadpool-postgres 0.14,
   tokio-postgres 0.7, postgres-types 0.2, pgvector 0.4, versions match
   ironclaw exactly) and [features] postgres = ["dep:..."]. default
   stays empty.

4. cap lib.rs: pub mod error; (always) +
   #[cfg(feature = "postgres")] pub mod repository;.

5. ironclaw Cargo.toml [features] postgres gains
   "dasclaw_workspace_cap/postgres" so the cap crate's postgres
   feature follows the workspace flag.

Validation (ALL pass):
- cargo check -p dasclaw_workspace_cap --tests
- cargo check -p dasclaw_workspace_cap --features postgres --tests
- cargo check -p dasclaw --tests
- cargo check -p dasclaw --no-default-features --features libsql,html-to-markdown --tests
- cargo nextest run -p dasclaw_workspace_cap --features postgres -> 113/113 passed
- cargo fmt --all
- python3.12 scripts/check_no_panics.py --base origin/xClaw
- cargo clippy --no-deps -p dasclaw_workspace_cap --all-targets -- -D warnings
- cargo clippy --no-deps -p dasclaw_workspace_cap --all-targets --features postgres -- -D warnings

Refs ADR-129 §1.3 (verbatim), ADR-152 §3 F3.4, ADR-114 (no new .ironclaw).
No new unwrap / expect / panic./cli/install-skills.sh --ai copilot No business-logic changes.

Closes #699
Refs #674
