+++ crates/dasclaw_workspace_cap/src/lib.rs
@@ -51,6 +51,11 @@ pub mod layer;
 pub mod policy;
 pub mod privacy;
 pub mod search;
+
+pub mod error;
+
+#[cfg(feature = "postgres")]
+pub mod repository;
 pub use policy::{
     NetworkAccess, SandboxPolicy, WritableRoot, default_read_only_subpaths_for_writable_root,
 };
